feat: Enable content tabs
This commit enables the pymdownx.tabbed extension in mkdocs.yml by uncommenting
its configuration. The `alternate_style:` true setting is included to provide an alternative visual
style for the tabs. This change allows for the creation of content tabs, which helps organize and
present related information concisely within the documentation.